fix: payment status update and mercado pago notification model serialization

+++ Infraestructure/Repositories/Classes/ConsultaRepository.cs
@@ -104,7 +104,7 @@ namespace Infraestructure.Repositories.Classes
             return appointmentAtDatabase;
         }
 
-        public async Task<Consulta> UpdatePaymentStatus(int id, string status)
+        public async Task<Consulta> UpdatePaymentStatus(long id, string status)
         {
             Consulta consultaAtDatabase = await _context.Consultas.FirstOrDefaultAsync(e => e.Id == id);
