+++ Infraestructure/Repositories/Interfaces/IConsultaRepository.cs
@@ -9,7 +9,7 @@ namespace Infraestructure.Repositories.Interfaces
         Task<Consulta> GetByPaymentId(int id);
         Task<Consulta> SetAppointmentAsCompleted(int id);
         Task<Consulta> SetPaymentId(int id, long paymentId);
-        Task<Consulta> UpdatePaymentStatus(int id, string status);
+        Task<Consulta> UpdatePaymentStatus(long id, string status);
         Task<IEnumerable<Consulta>?> GetAllPatientAppointmentsScheduled(int id);
         Task<IEnumerable<Consulta>?> GetAllPatientAppointmentsCompleted(int id);
         Task<IEnumerable<Consulta>?> GetAllDoctorAppointmentsScheduled(int id);
